G-002_GateOfBabylon/eastmoney/requirements.txt
httpx
psycopg2-binary
pydantic
PyYAML
